requirements.txt
PyQt6>=6.6.1
google-generativeai>=0.3.2
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
mss>=9.0.1
pytesseract>=0.3.10
PyMuPDF>=1.23.8
python-docx>=1.1.0
python-dotenv>=1.0.0
keyboard>=0.13.5
pywin32>=306 